third_party/blink/web_tests/crypto/subtle/pbkdf2/deriveBits-expected.txt

Tests importKey/deriveBits for PBKDF2

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [a7950c143ec64e2b8d4bb1db8677188b] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [7deaf8b4a801011c1cd27f36e3bfc962] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [83eb100b6a3a975f0fe3ffcdc2419852] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [e4dfce3830983830c50c351a0b0f79e1] and was
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "PBKDF2"
PASS tmpKey.usages.join(',') is "deriveKey,deriveBits"
PASS: deriveBits should be [750261780a187897a9978371599db5d1] and was
PASS successfullyParsed is true

TEST COMPLETE
